requirements.txt
pandas==2.2.1
openpyxl==3.1.2
requests==2.31.0
python-dateutil==2.8.2
schedule==1.2.1
python-dotenv==1.0.1
urllib3==2.1.0
certifi>=2024.2.2
pytz>=2024.1
SQLAlchemy>=2.0.0
pyodbc>=5.1.0
cryptography>=42.0.0
numpy>=1.26.0
